Add files via upload
A file manager is a software program that provides a graphical user interface (GUI) for organizing and managing files and folders on a computer. It allows users to view, copy, move, rename, delete, and search for files and folders. File managers typically display files and folders in a hierarchical tree structure, making it easy for users to navigate through their file system. They often include features such as file previews, file properties, and the ability to create and extract compressed files. File managers are essential tools for managing and organizing files on a computer efficiently.